ARM: Use -mcpu=cortex-a15 when targeting MSM Krait CPUs
Enable compiler optimizations specific to the Cortex-A15 processor
when targeting MSM Krait CPUs. This is necessary take advantage of
the UDIV/SDIV instructions supported by these processors.  To
accomplish this, we need to remove the -march=armv7-a ISA restriction
from the compiler options because 'cortex-a15' is a superset of 'armv7-a'.

Change-Id: I6215aecc11fb4f77c971de7b84f68649ef234357
Signed-off-by: Stepan Moskovchenko <stepanm@codeaurora.org>