ASoC: wcd9xxx: Introduce IOCTL_BYPASS flag
It might be useful to allow the requests to write that come from IOCTL to
actually do the write. In case there is an IOCTL filter in sound codec, it
is not possible to monitor the handled registers' values changes which are
triggered by the userspace calls via IOCTL as these changes get filtered
out by design.  A filter should be ommited to acknowledge the exact gain
values that are being set by Android AudioHAL.

With a new IOCTL_BYPASS flag it is easy to switch between SysFS/IOCTL and
SysFS only write modes, therefore it is not required to recompile the
kernel with filter dropped from sound codec to allow both the writes from
IOCTL and reads via a Sound Gain Controller with a presence of the write
call in control data.

Note that the sound codec that uses control data must utilize this flag in
its filter call. Otherwise there will be no sence in using it.

Change-Id: Id216eed75680dd9192b89d24c1a5eef55efe7f94
Signed-off-by: Alex Saiko <solcmdr@gmail.com>